+++ index.html
@@ -1,4 +1,4 @@
-<!DOCTYPE html>
+<!doctype html>
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
@@ -13,7 +13,7 @@
     <script type="module" src="./src/js/main.js" defer></script>
     <script
       type="module"
-      src="./src/js/listings/fetchListings.mjs"
+      src="./src/js/listings/displayListings.mjs"
       defer
     ></script>
     <script
